requirements.txt
discord.py==1.5.1
wikipedia==1.4.0
praw==7.1.0